feat(mobile): migrate to iOS native liquid glass UI with NativeTabs
Replace JS Tabs with NativeTabs (expo-router/unstable-native-tabs) for
native liquid glass tab bar on iOS 26+. Restructure routes into nested
groups (share/request) with Slot layouts for seamless tab switching.
Add GlassView cards, SF Symbols, PlatformColor, pull-to-refresh, and
module-level caching to prevent reload flicker on tab switch.
