+++ apps/mobile/app/(tabs)/(request)/_layout.tsx
@@ -1,0 +1,5 @@
+import { Slot } from "expo-router";
+
+export default function RequestLayout() {
+  return <Slot />;
+}
