+++ apps/mobile/app/auth/sign-in.tsx
@@ -1,5 +1,6 @@
-import { View, Text, StyleSheet, TouchableOpacity, Alert } from 'react-native';
-import { SafeAreaView } from 'react-native-safe-area-context';
+import { View, Text, Pressable, Alert, PlatformColor, ActivityIndicator } from 'react-native';
+import { GlassView, isLiquidGlassAvailable } from 'expo-glass-effect';
+import { SymbolView } from 'expo-symbols';
 import { supabase } from '../../../../lib/supabase';
 import * as WebBrowser from 'expo-web-browser';
 import * as Linking from 'expo-linking';
@@ -13,7 +14,6 @@ export default function SignIn() {
   const handleGoogleSignIn = async () => {
     setLoading(true);
     try {
-      // Use makeUrl for Expo Go compatibility, createURL for dev builds
       const redirectUrl = Linking.createURL('auth/callback');
       console.log('Redirect URL:', redirectUrl);
 
@@ -23,7 +23,6 @@ export default function SignIn() {
           redirectTo: redirectUrl,
           skipBrowserRedirect: true,
           queryParams: {
-            // Force Google to show account picker every time (don't reuse last account after logout)
             prompt: 'select_account',
           },
         },
@@ -43,13 +42,11 @@ export default function SignIn() {
 
       if (result.type === 'success') {
         const { url } = result;
-        // Extract the URL params
         const urlParams = new URLSearchParams(url.split('#')[1] || url.split('?')[1]);
         const accessToken = urlParams.get('access_token');
         const refreshToken = urlParams.get('refresh_token');
 
         if (accessToken && refreshToken) {
-          // Set the session with the tokens
           await supabase.auth.setSession({
             access_token: accessToken,
             refresh_token: refreshToken,
@@ -64,58 +61,94 @@ export default function SignIn() {
   };
 
   return (
-    <SafeAreaView style={styles.container}>
-      <View style={styles.content}>
-        <Text style={styles.title}>SlugSwap</Text>
-        <Text style={styles.subtitle}>Share dining points with fellow students</Text>
+    <View style={{
+      flex: 1,
+      justifyContent: 'center',
+      alignItems: 'center',
+      padding: 24,
+      backgroundColor: PlatformColor('systemGroupedBackground'),
+    }}>
+      <View style={{ alignItems: 'center', marginBottom: 48 }}>
+        <SymbolView
+          name="fork.knife.circle.fill"
+          tintColor={PlatformColor('systemBlue')}
+          size={64}
+          weight="light"
+        />
+        <Text style={{
+          fontSize: 42,
+          fontWeight: 'bold',
+          color: PlatformColor('label'),
+          marginTop: 16,
+        }}>
+          SlugSwap
+        </Text>
+        <Text style={{
+          fontSize: 17,
+          color: PlatformColor('secondaryLabel'),
+          marginTop: 8,
+        }}>
+          Share dining points with fellow students
+        </Text>
+      </View>
 
-        <TouchableOpacity
-          style={[styles.button, loading && styles.buttonDisabled]}
+      {isLiquidGlassAvailable() ? (
+        <GlassView isInteractive style={{ borderRadius: 14, width: '100%' }}>
+          <Pressable
+            onPress={handleGoogleSignIn}
+            disabled={loading}
+            style={({ pressed }) => ({
+              flexDirection: 'row',
+              alignItems: 'center',
+              justifyContent: 'center',
+              gap: 8,
+              paddingVertical: 16,
+              paddingHorizontal: 32,
+              opacity: pressed ? 0.7 : loading ? 0.6 : 1,
+            })}
+          >
+            {loading ? (
+              <ActivityIndicator color={PlatformColor('label')} />
+            ) : (
+              <>
+                <SymbolView name="person.crop.circle" tintColor={PlatformColor('label')} size={20} />
+                <Text style={{ color: PlatformColor('label'), fontSize: 17, fontWeight: '600' }}>
+                  Sign in with Google
+                </Text>
+              </>
+            )}
+          </Pressable>
+        </GlassView>
+      ) : (
+        <Pressable
           onPress={handleGoogleSignIn}
           disabled={loading}
+          style={({ pressed }) => ({
+            flexDirection: 'row',
+            alignItems: 'center',
+            justifyContent: 'center',
+            gap: 8,
+            paddingVertical: 16,
+            paddingHorizontal: 32,
+            borderRadius: 14,
+            borderCurve: 'continuous',
+            backgroundColor: PlatformColor('systemBlue'),
+            opacity: pressed ? 0.7 : loading ? 0.6 : 1,
+            width: '100%',
+          })}
         >
-          <Text style={styles.buttonText}>
-            {loading ? 'Signing in...' : 'Sign in with Google'}
-          </Text>
-        </TouchableOpacity>
-      </View>
-    </SafeAreaView>
+          {loading ? (
+            <ActivityIndicator color="#fff" />
+          ) : (
+            <>
+              <SymbolView name="person.crop.circle" tintColor="#fff" size={20} />
+              <Text style={{ color: '#fff', fontSize: 17, fontWeight: '600' }}>
+                Sign in with Google
+              </Text>
+            </>
+          )}
+        </Pressable>
+      )}
+    </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-  },
-  content: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 24,
-  },
-  title: {
-    fontSize: 48,
-    fontWeight: 'bold',
-    marginBottom: 8,
-  },
-  subtitle: {
-    fontSize: 16,
-    color: '#666',
-    marginBottom: 48,
-  },
-  button: {
-    backgroundColor: '#007AFF',
-    paddingHorizontal: 32,
-    paddingVertical: 16,
-    borderRadius: 8,
-  },
-  buttonDisabled: {
-    opacity: 0.6,
-  },
-  buttonText: {
-    color: '#fff',
-    fontSize: 16,
-    fontWeight: '600',
-  },
-});
