+++ apps/mobile/package.json
@@ -16,11 +16,14 @@
     "@supabase/supabase-js": "^2.95.3",
     "expo": "~54.0.33",
     "expo-barcode-pdf417": "^1.0.0",
+    "expo-blur": "^15.0.8",
     "expo-constants": "^18.0.13",
     "expo-dev-client": "~6.0.20",
+    "expo-glass-effect": "^0.1.9",
     "expo-linking": "^8.0.11",
     "expo-router": "^6.0.23",
     "expo-status-bar": "~3.0.9",
+    "expo-symbols": "^1.0.8",
     "expo-web-browser": "^15.0.10",
     "react": "19.1.0",
     "react-native": "0.81.5",
