+++ package.json
@@ -47,7 +47,9 @@
     "next": "^15.5.6",
     "react": "19.1.0",
     "react-dom": "19.1.0",
-    "vercel": "^50.14.0"
+    "vercel": "^50.14.0",
+    "expo": "~54.0.33",
+    "react-native": "0.81.5"
   },
   "devDependencies": {
     "@types/cors": "^2.8.19",
